feat: add search functionality and updated code to dynamically update dashboard

+++ src/app/(components)/pet-carriers/actions.ts
@@ -19,6 +19,22 @@ export const getErrorMessage = (error: unknown): string => {
   return message;
 };
 
+export const fetchRecords = async (search?: string) => {
+  const petCarriersData = await prisma.petCarriers.findMany({
+    where: {
+      carrierName: {
+        contains: search,
+        mode: "insensitive",
+      },
+    },
+    orderBy: {
+      createdAt: "asc",
+    },
+  });
+
+  return petCarriersData;
+};
+
 export const createRecord = async (formData: FormData) => {
   const newName = formData.get("carrier-name") as string;
 
@@ -26,6 +42,7 @@ export const createRecord = async (formData: FormData) => {
     await prisma.petCarriers.create({
       data: {
         carrierName: newName,
+        updatedAt: null,
       },
     });
   } catch (error) {
@@ -63,6 +80,13 @@ export const deleteRecord = async (formData: FormData) => {
   const id = formData.get("carrier-id") as string;
 
   try {
+    const rentalCount = await prisma.rentals.count({
+      where: { carrierId: id },
+    });
+    if (rentalCount > 0) {
+      return { error: "Delete linked rental record first" };
+    }
+
     await prisma.petCarriers.delete({
       where: { carrierId: id },
     });
@@ -74,13 +98,3 @@ export const deleteRecord = async (formData: FormData) => {
 
   revalidatePath("/pet-carriers");
 };
-
-// export const getCarrier = async (formData: FormData) => {
-//   const carrier = formData.get("carrier-search") as string;
-
-//   await prisma.petCarriers.findUnique({
-//     where: {
-//       carrierId: carrier,
-//     },
-//   });
-// };
